Refactored Generator flow and updated NavigationController
