+++ Runar/Runar/Controllers/Generator/SelectRune/SelectRuneCollectionView.swift
@@ -23,7 +23,7 @@ public class SelectRuneCollectionView: UICollectionView, UICollectionViewDataSou
         fatalError("init(coder:) has not been implemented")
     }
 
-    private func setupViews(){
+    private func setupViews() {
         backgroundColor = .clear
         
         dataSource = self
@@ -71,7 +71,7 @@ public class SelectRuneCollectionView: UICollectionView, UICollectionViewDataSou
         self.selectedRunesCount -= 1
     }
     
-    @IBAction func selectRune(runeImage: UIButton){
+    @objc func selectRune(runeImage: UIButton) {
         if (self.selectedRunesCount < 3) {
             let cell = runeImage.superview as! SelectRuneCell
             
@@ -86,62 +86,5 @@ public class SelectRuneCollectionView: UICollectionView, UICollectionViewDataSou
     
     func getRune(at index: Int) -> SelectRuneCell {
         return self.runes[index]
-    }
-}
-
-class SelectRuneCell: UICollectionViewCell {
-    public var model: GenerationRuneModel?
-    public var indexPath: IndexPath = IndexPath()
-    
-    override public init(frame: CGRect) {
-        super.init(frame: frame)
-        setupViews()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    let runeImage: UIButton = {
-        let runeImage = UIButton()
-        
-        runeImage.contentMode = .scaleAspectFill
-        
-        return runeImage
-    }()
-    
-    private func setupViews(){
-        backgroundColor = .clear
-        
-        addSubview(runeImage)
-        
-        runeImage.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            runeImage.heightAnchor.constraint(equalToConstant: 97), //78
-            runeImage.widthAnchor.constraint(equalToConstant: 78), //66
-            runeImage.centerXAnchor.constraint(equalTo: centerXAnchor),
-            runeImage.centerYAnchor.constraint(equalTo: centerYAnchor)
-        ])
-    }
-    
-    public func setRune(_ rune: GenerationRuneModel, _ indexPath: IndexPath){
-        self.model = rune
-        self.indexPath = indexPath
-        
-        runeImage.setBackgroundImage(rune.image.image, for: .normal)
-    }
-    
-    public func selectRune() {
-        self.isSelected = true
-        self.toggleState()
-    }
-    
-    public func deselectRune() {
-        self.isSelected = false
-        self.toggleState()
-    }
-    
-    public func toggleState(){
-        self.runeImage.isEnabled = !self.isSelected
     }
 }
