+++ Runar/Runar/Extensions/Array + Extensions.swift
@@ -1,4 +1,6 @@
-public extension Array {
+import Foundation
+
+extension Array {
     subscript(safe index: Int) -> Element? {
         0..<self.count ~= index ? self[index]: nil
     }
