+++ Runar/Runar/Extensions/UIView.swift
@@ -5,9 +5,10 @@
 //  Created by Юлия Лопатина on 12.04.21.
 //
 
-import UIKit.UIView
+import UIKit
 
 extension UIView {
+
     func addSubviews(_ views: UIView...) {
         for view in views {
             addSubview(view)
